\name{message.queue}
\Rdversion{1.1}
\docType{class}
\alias{message.queue-class}
\alias{message.queue}

\title{Class ``message.queue''}
\description{The \code{message.queue} is derived from the 
\code{queue.interface}.  Message queues provide interprocess communication
on a single computer.
}
\keyword{classes}
